requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.2
python-dotenv==1.0.0
Werkzeug==2.3.7
email-validator==2.1.0.post1
Flask-WTF==1.2.1
Flask-Migrate==4.0.5
SQLAlchemy==2.0.23
python-dateutil==2.8.2
